requirements.txt
selenium
tqdm
requests
lxml
python-decouple
